+++ src/Sentry/Laravel/ServiceProvider.php
@@ -7,12 +7,14 @@ use Illuminate\Foundation\Application as Laravel;
 use Illuminate\Foundation\Http\Kernel as HttpKernel;
 use Illuminate\Log\LogManager;
 use Laravel\Lumen\Application as Lumen;
+use RuntimeException;
 use Sentry\ClientBuilder;
 use Sentry\ClientBuilderInterface;
 use Sentry\Integration as SdkIntegration;
 use Sentry\Laravel\Http\LaravelRequestFetcher;
 use Sentry\Laravel\Http\SetRequestIpMiddleware;
 use Sentry\Laravel\Http\SetRequestMiddleware;
+use Sentry\Laravel\Tracing\ServiceProvider as TracingServiceProvider;
 use Sentry\SentrySdk;
 use Sentry\State\Hub;
 use Sentry\State\HubInterface;
@@ -23,7 +25,7 @@ class ServiceProvider extends BaseServiceProvider
      * List of configuration options that are Laravel specific and should not be sent to the base PHP SDK.
      */
     private const LARAVEL_SPECIFIC_OPTIONS = [
-        // We do not want these settings to hit the PHP SDK because it's Laravel specific and the SDK will throw errors
+        // We do not want these settings to hit the PHP SDK because they are Laravel specific and the PHP SDK will throw errors
         'tracing',
         'breadcrumbs',
         // We resolve the integrations through the container later, so we initially do not pass it to the SDK yet
@@ -132,7 +134,7 @@ class ServiceProvider extends BaseServiceProvider
         }
 
         $this->app->bind(ClientBuilderInterface::class, function () {
-            $basePath = base_path();
+            $basePath   = base_path();
             $userConfig = $this->getUserConfig();
 
             foreach (self::LARAVEL_SPECIFIC_OPTIONS as $laravelSpecificOptionName) {
@@ -141,7 +143,7 @@ class ServiceProvider extends BaseServiceProvider
 
             $options = \array_merge(
                 [
-                    'prefixes' => [$basePath],
+                    'prefixes'       => [$basePath],
                     'in_app_exclude' => ["{$basePath}/vendor"],
                 ],
                 $userConfig
@@ -228,21 +230,40 @@ class ServiceProvider extends BaseServiceProvider
             new Integration\ExceptionContextIntegration,
         ];
 
-        $userIntegrations = $this->getUserConfig()['integrations'] ?? [];
+        $userConfig = $this->getUserConfig();
 
-        foreach ($userIntegrations as $userIntegration) {
+        $integrationsToResolve = $userConfig['integrations'] ?? [];
+
+        $enableDefaultTracingIntegrations = $userConfig['tracing']['default_integrations'] ?? true;
+
+        if ($enableDefaultTracingIntegrations && $this->couldHavePerformanceTracingEnabled()) {
+            $integrationsToResolve = array_merge($integrationsToResolve, TracingServiceProvider::DEFAULT_INTEGRATIONS);
+        }
+
+        foreach ($integrationsToResolve as $userIntegration) {
             if ($userIntegration instanceof SdkIntegration\IntegrationInterface) {
                 $integrations[] = $userIntegration;
             } elseif (\is_string($userIntegration)) {
                 $resolvedIntegration = $this->app->make($userIntegration);
 
-                if (!($resolvedIntegration instanceof SdkIntegration\IntegrationInterface)) {
-                    throw new \RuntimeException('Sentry integrations should a instance of `\Sentry\Integration\IntegrationInterface`.');
+                if (!$resolvedIntegration instanceof SdkIntegration\IntegrationInterface) {
+                    throw new RuntimeException(
+                        sprintf(
+                            'Sentry integrations must be an instance of `%s` got `%s`.',
+                            SdkIntegration\IntegrationInterface::class,
+                            get_class($resolvedIntegration)
+                        )
+                    );
                 }
 
                 $integrations[] = $resolvedIntegration;
             } else {
-                throw new \RuntimeException('Sentry integrations should either be a container reference or a instance of `\Sentry\Integration\IntegrationInterface`.');
+                throw new RuntimeException(
+                    sprintf(
+                        'Sentry integrations must either be a valid container reference or an instance of `%s`.',
+                        SdkIntegration\IntegrationInterface::class
+                    )
+                );
             }
         }